requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
openai==1.3.7
sentence-transformers==2.2.2
chromadb==0.4.18
Pillow==10.1.0
PyMuPDF==1.23.8
pdf2image==1.16.3
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
huggingface-hub==0.19.4
torch==2.1.1
torchvision==0.16.1
transformers==4.36.0
numpy==1.24.3
matplotlib==3.7.2
mineru==0.2.0